Move header component
